restructure the dashboard in seperate modules
